requirements.txt
openai~=1.55.0
python-dotenv==1.0.1
pyaudio==0.2.14